Move the adapter into separate sonar subpackage
