+++ src/test/java/sorald/SegmentStrategyTest.java
@@ -5,7 +5,7 @@ import org.sonar.java.checks.ArrayHashCodeAndToStringCheck;
 import sorald.Constants;
 import sorald.Main;
 import sorald.PrettyPrintingStrategy;
-import sorald.SonarVerifierAdapter;
+import sorald.sonar.SonarVerifierAdapter;
 import sorald.TestHelper;
 
 public class SegmentStrategyTest {
